+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/v2/orc/OrcDataSourceV2.scala
@@ -17,7 +17,7 @@
 package org.apache.spark.sql.execution.datasources.v2.orc
 
 import org.apache.spark.sql.execution.datasources._
-import org.apache.spark.sql.execution.datasources.orc._
+import org.apache.spark.sql.execution.datasources.orc.OrcFileFormat
 import org.apache.spark.sql.execution.datasources.v2._
 import org.apache.spark.sql.sources.v2.{DataSourceOptions, Table}
 import org.apache.spark.sql.types.StructType
